% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigsVisualization.R
\name{ContigsVisualization}
\alias{ContigsVisualization}
\title{CircosVisualization}
\usage{
ContigsVisualization(fna_count = NULL, completeness = FALSE, gb_dir = NULL)
}
\arguments{
\item{fna_count}{A completeness of genome. If NULL, all nucleotides are recognized as chromosome driven.}

\item{completeness}{A completeness of genome. If NULL, all nucleotides are recognized as chromosome driven.}

\item{gb_dir}{A completeness of genome. If NULL, all nucleotides are recognized as chromosome driven.}
}
\value{
A genome map.
}
\description{
CircosVisualization
}
