% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_ratio.R
\name{gc_ratio}
\alias{gc_ratio}
\title{gc_ratio #}
\usage{
gc_ratio(x)
}
\arguments{
\item{sequence}{A nucleotide sequence.}
}
\value{
A list containing summary information of the processed data.
}
\description{
gc_ratio #
}
