% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_skew.R
\name{gc_skew}
\alias{gc_skew}
\title{gc_skew #}
\usage{
gc_skew(x)
}
\arguments{
\item{sequence}{A nucleotide sequence.}
}
\value{
A list containing summary information of the processed data.
}
\description{
gc_skew #
}
