% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDrawer.R
\name{GenomeDrawer}
\alias{GenomeDrawer}
\title{GenomeDrawer #}
\usage{
GenomeDrawer(dir = NULL)
}
\arguments{
\item{dir}{A string specifying the directory where the gff output files are located. If NULL, the current working directory is used.}
}
\value{
A list containing information for plotting genome map.
}
\description{
GenomeDrawer #
}
