% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeVisualization.R
\name{GenomeVisualization}
\alias{GenomeVisualization}
\title{CircosVisualization}
\usage{
GenomeVisualization(fna_count = NULL, completeness = TRUE)
}
\arguments{
\item{completeness}{A completeness of genome. If NULL, all nucleotides are recognized as chromosome driven.}

\item{i}{A number of contigs.}
}
\value{
A genome map.
}
\description{
CircosVisualization
}
