% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDrawer.R
\name{run_drawer}
\alias{run_drawer}
\title{RunDrawer}
\usage{
run_drawer(completeness = NULL)
}
\arguments{
\item{completness}{A genbank completeness of data.}
}
\value{
A genome map.
}
\description{
RunDrawer
}
